Fix compilation with multiple usage of magneto library
